Revert "PRS-26_API /dropdownsearch erstellen"
