+++ gunicorn.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-gunicorn --chdir app app:app -w 2 -b 0.0.0.0:80
+gunicorn --chdir app main:app -w 2 -b 0.0.0.0:80
